Modified batched residuals return of torch.linalg.lstsq
SciPy and NumPy operate only on non-batched input and return an
empty array with shape (0,) if rank(a) != n.
The behavior for non-batched inputs is NumPy and SciPy compatible and the
same result is computed.
For batched inputs, if any matrix in the batch has rank less than `n`,
then an empty tensor is returned.

ghstack-source-id: 2221373c23de5a78f19b7df0a7a381f611647bee
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54722